+++ tests/component/phases.spec.ts
@@ -1,6 +1,5 @@
 import {generatePhaseTests} from '../_data/lifecycle/tests';
-import {ComponentPhase} from '../../src/component/phase';
-import {componentPhase} from '../../src/component/phase';
+import {type ComponentPhase, componentPhase} from '../../src/component/phase';
 import {componentPhases} from '../../src/component/phases';
 import {SampleComponent} from '../_data/sample/component';