+++ tests/invoke/listener.spec.ts
@@ -126,7 +126,10 @@ describe('invokeListener', () => {
 			expect(spyA).not.toHaveBeenCalled();
 			expect(spyB).not.toHaveBeenCalled();
 
-			const _result = await invokeListeners<ServerPhase, SampleServer>('serverDidLoad', custom);
+			const _result = await invokeListeners<ServerPhase, SampleServer>({
+				phase: 'serverDidLoad',
+				delegate: custom
+			});
 
 			expect(spyA).toHaveBeenCalledTimes(1);
 			expect(spyB).toHaveBeenCalledTimes(1);